requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None
#
aiofiles==25.1.0
    # via webscout
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via webscout
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
beautifulsoup4==4.14.2
    # via markdownify
brotli==1.1.0
    # via webscout
bson==0.5.10
    # via webscout
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cloudscraper==1.2.71
    # via webscout
colorama==0.4.6
    # via webscout
curl-cffi==0.13.0
    # via webscout
deprecated==1.2.18
    # via nodriver
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
environs==14.3.0
    # via -r requirements.in
fake-useragent==2.2.0
    # via -r requirements.in
fastapi[standard]==0.119.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.14
    # via fastapi
fastapi-cloud-cli==0.3.1
    # via fastapi-cli
filelock==3.20.0
    # via huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.9.0
    # via
    #   gradio-client
    #   huggingface-hub
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.26.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.185.0
    # via google-generativeai
google-auth==2.41.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via webscout
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
gradio-client==1.13.3
    # via webscout
grpcio==1.75.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.10
    # via huggingface-hub
html5lib==1.1
    # via webscout
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   gradio-client
    #   openai
huggingface-hub==0.35.3
    # via gradio-client
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via wasmtime
jinja2==3.1.6
    # via fastapi
jiter==0.11.1
    # via openai
litproxy==0.2.1
    # via webscout
lxml==6.0.2
    # via webscout
markdown-it-py==4.0.0
    # via rich
markdownify==1.2.0
    # via webscout
markupsafe==3.0.3
    # via jinja2
marshmallow==4.0.1
    # via environs
mdurl==0.1.2
    # via markdown-it-py
mistune==3.1.4
    # via webscout
mouseinfo==0.1.3
    # via pyautogui
mss==10.1.0
    # via nodriver
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via webscout
nodriver==0.47.0
    # via webscout
numpy==2.3.4
    # via -r requirements.in
openai==2.6.0
    # via webscout
orjson==3.11.3
    # via webscout
packaging==25.0
    # via
    #   gradio-client
    #   huggingface-hub
pillow==12.0.0
    # via
    #   -r requirements.in
    #   pyscreeze
    #   webscout
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psutil==7.1.1
    # via webscout
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyautogui==0.9.54
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic[email]==2.12.3
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cloud-cli
    #   google-generativeai
    #   openai
pydantic-core==2.41.4
    # via pydantic
pygetwindow==0.0.9
    # via pyautogui
pygments==2.19.2
    # via rich
pymsgbox==2.0.1
    # via pyautogui
pyparsing==3.2.5
    # via
    #   cloudscraper
    #   httplib2
pyperclip==1.11.0
    # via mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
python-dateutil==2.9.0.post0
    # via bson
python-dotenv==1.1.1
    # via
    #   environs
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
python3-xlib==0.15
    # via
    #   mouseinfo
    #   pyautogui
pytweening==1.2.0
    # via pyautogui
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   uvicorn
    #   webscout
requests==2.32.5
    # via
    #   cloudscraper
    #   google-api-core
    #   huggingface-hub
    #   litproxy
    #   requests-toolbelt
    #   webscout
requests-toolbelt==1.0.0
    # via cloudscraper
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
    #   webscout
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.1
    # via fastapi-cloud-cli
rsa==4.9.1
    # via google-auth
ruamel-yaml==0.18.15
    # via -r requirements.in
ruamel-yaml-clib==0.2.14
    # via ruamel-yaml
sentry-sdk==2.42.1
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   bson
    #   html5lib
    #   markdownify
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
starlette==0.48.0
    # via fastapi
tls-client==1.0.1
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   google-generativeai
    #   gradio-client
    #   grpcio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.38.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
wasmtime==38.0.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
webencodings==0.5.1
    # via html5lib
webscout==8.3.7
    # via -r requirements.in
websockets==15.0.1
    # via
    #   gradio-client
    #   nodriver
    #   uvicorn
wheel==0.45.1
    # via webscout
wrapt==1.17.3
    # via deprecated
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via webscout

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
